Update to auto-size height
Changed height from 1000 to auto so that the height of the player sheet is only as tall as the content contained within it. This is so changes to the HTML to only show limited information when actor is Limited permission won't still be 1000px height when there's only a third of that space being used.